Now using long instead of BigInteger.

+++ src/main/java/com/github/open96/primenumbergenerator/sieve/Sieve.java
@@ -8,5 +8,5 @@ import java.math.BigInteger;
 public interface Sieve {
     void deleteNonPrimeNumbers();
     void printSieve();
-    boolean checkIfNumberIsPrime(BigInteger number);
+    boolean checkIfNumberIsPrime(long number);
 }
